Attempt #2 of Exercism.io Resistor-color-duo JavaScript Challenge
Mentor recommendation: 
Instead of coercing the numbers returned from Array.indexOf to be strings, combining those two strings, and then coercing the resulting string to be a number, there is a way you could combine those two numbers into your return value. As a hint, don't think in terms of concatenation; think of it as an arithmetic problem. How would you combine two numbers into a single number, one representing the 'ones' value and the other representing the 'tens' portion of the final value?